feat: openapi http-api code generator
