+++ examples/micronaut-app/src/test/kotlin/ht/eyfout/pets/CodeGeneratedHttpAPITests.kt
@@ -3,7 +3,7 @@ package ht.eyfout.pets
 import ht.eyfout.example.client.pets.PetClient
 import ht.eyfout.example.controller.PetsController
 import ht.eyfout.junit.jupiter.gherkin.api.GherkinDynamicTest
-import ht.eyfout.junit.jupiter.gherkin.http.generated.warehouse.GETloginUserHttpAPI
+import ht.eyfout.openapi.http.api.generated.warehouse.GETloginUserHttpAPI
 import ht.eyfout.pets.http.PetsStateScopeProvider
 import io.micronaut.http.HttpResponse
 import io.micronaut.test.annotation.MockBean
