--- libraries/gherkin-http-api-generator/src/main/java/ht/eyfout/junit/jupiter/gherkin/http/GherkinHttpAPIGenerationException.java
@@ -1,7 +1,0 @@
-package ht.eyfout.junit.jupiter.gherkin.http;
-
-public class GherkinHttpAPIGenerationException extends RuntimeException{
-    public GherkinHttpAPIGenerationException(String msg, Throwable t){
-        super(msg, t);
-    }
-}
