--- libraries/gherkin-http-api-generator/src/main/java/ht/eyfout/junit/jupiter/gherkin/http/Pair.java
@@ -1,4 +1,0 @@
-package ht.eyfout.junit.jupiter.gherkin.http;
-
-public record Pair<F, S>(F first, S second) {
-}
