+++ libraries/http-api-provider/build.gradle
