+++ libraries/openapi-http-api-generator/src/main/java/ht/eyfout/openapi/http/api/GherkinHttpAPIGenerationException.java
@@ -1,0 +1,7 @@
+package ht.eyfout.openapi.http.api;
+
+public class GherkinHttpAPIGenerationException extends RuntimeException {
+    public GherkinHttpAPIGenerationException(String msg, Throwable t) {
+        super(msg, t);
+    }
+}
