+++ libraries/openapi-http-api-generator/src/main/java/ht/eyfout/openapi/http/api/Pair.java
@@ -1,0 +1,4 @@
+package ht.eyfout.openapi.http.api;
+
+public record Pair<F, S>(F first, S second) {
+}
