+++ libraries/openapi-http-api-generator/src/test/resources/pet-store.yml
