+++ libraries/openapi-http-api-generator/src/test/resources/supply-warehouse.yml
